Fix secretmigrator condition and tests
Use the correct condition and Do* function for the new RKE field
migration. We don't need to explicitly set the condition to True when
DoUntilTrue is used.

Fix the unit tests to check that the object doesn't change when the
migration is recalled, to show that the conditions aren't changed every
controller sync.

(cherry picked from commit b233d708aea8dea96ad7d893c620682a7bfa0014)