Add API document and output test data to it
This change updates the Steve API tests to output sample data. The
example requests and responses are stored in a README. This provides
external users with validated documentation of all the Steve parameters.

When updating or adding tests, the integration tests must be run locally
in order to generate the samples. The subset of relevant tests can be
run with:

go test -count=1 -v ./tests/v2/integration/steveapi/

If any outputs have changed, update the README with

cd ./tests/v2/integration/steveapi/
./make-table.sh

Then commit the changes.