Fix the issue that PSACT is not restored on RKE1 clusters.
Previously, when we restore the snapshot on an RKE1 cluster and choose "restore cluster config, kubernetes version and etcd", the PSACT is not restored.
It happens because the field ".spec.defaultPodSecurityAdmissionConfigurationTemplateName" is not set back to the value from the backup.
This fix is to set the field to the value from the backup.