Convert RBAC handlers to use manager interfaces
The managementuser RBAC handlers are inconsistent in their use of the
listers and interfaces that are already stored and initialized on the
`manager` object versus the ones embedded on UserContext. This is
confusing to read and makes it difficult to mock out caches and clients
in unit tests. This change converts uses of `m.workload` to an
appropriate norman lister or interface where appropriate.