Changing webhook in remote to use the right image
The rancher webhook is now deployed in downstream clusters. This
caused an issue where, in airgap scenarios, the webhook and the
pods that were deploying the webhook would not use the specified
airgap registry. This change allows the agent to pass down an
override image, when applicable.

(cherry picked from commit 389e31e3421042a47d5d918d3708511fd1143b3e)