Add setup script for project resources API
The project resources API points to a Service named either "rancher" or
"cattle-cluster-agent", whichever it can find, at which the Kubernetes
API server can access the API. For helm and docker setups, there exists
a Service named "rancher" pointing to the rancher container or
Deployment. For the cluster agent, the service is named
"cattle-cluster-agent". For developer instances of Rancher running
directly on localhost, there is no automatic Service deployment. This
script allows the new API to work by creating a Service to forward
traffic from the API server to the local Rancher instance.

Other types of Service configurations, like using ExternalName, were
experimented with to avoid the extra pod, but none could successfully
bridge the traffic gap from the cluster to the Rancher instance.