Ensure cluster controllers are always reenqueued
Before 23d89f9, when a cluster entered the managementapi user
controller, it was guaranteed to enqueue all cluster controllers at the
end of the sync unless it encountered an error. With 23d89f9, the sync
would exit early in many cases, thereby failing to enqueue the cluster
controllers. When a cluster is deleted, one of these cluster controllers
is supposed to call Stop on the cluster manager and thereby stop the
context, but since this was not always happening, the context sometimes
remained running and would mean the healthsyncer controller continued to
run and interminably emit errors for the now missing cluster.

This change addresses the problem by adding calls to enqueue the cluster
controllers after every success return.

It also prevents reenqueuing nil clusters by copying the cluster objects
and never nil-ifying the clusters on errors.

To improve readability and debuggability, all errors returned from the
sync method are wrapped, and the private const `_all_` is replaced by
the imported `AllKey` from wrangler, to make it clear that this key has
meaning to wrangler and is not a random string.