Create project member binding for system user
Prior, the project member ProjectRoleTemplateBinding was created by
a project handler that was deferred on the pipeline type. This means
that the one handler which created the binding only executed if
pipelines were assigned to the downstream cluster. Pipelines are now
deprecated, making it less likely this condition will be met. Project
monitoring depends on this binding. Now, the project member
ProjectRoleTemplateBinding is additionally created in a project
handler that is only deferred until cluster monitoring is enabled.
The handler will then proceed to create the binding if monitoring is
enabled for the given project.