Run steveapi tests on local and downstream
The behavior on local vs downstream can sometimes be subtly different,
so it is worth running API tests on both. Also, the steveapi tests are
self-documenting, but they were not writing their outputs because their
outputs were different from run to run because the downstream cluster ID
was different. This change ensures both sets of tests are run but the
output examples are only recorded when the output is deterministic
(which is for the local).