Make steveapi tests more flexible
The steveapi integration tests try to be precise about the desired
output of API calls. This was causing a problem when run on k8s 1.23
which has default token secrets in every namespace and so was adding
extra resources to the response output. This change adds a label to the
secrets created in TestList so that they can easily be identified on a
live cluster and so unlabeled secrets can be ignored.