Initial round of unit tests for Okta+LDAP
These are specifically intended to test the behavior in PR #41269
so they are intentionally quite limited in scope. Mostly the goal
is to ensure that when an ldapProvider is configured on a SAML
provider, it is actually used when a principal search is performed.

This would be fairly trivial to expand to the shibboleth provider,
and in the future I'd like to include a group search suite.