Use a random ID for integration test labels
5f78652 introduced labels to resources in the steveapi integration tests
so that assertions could exclude non-test-generated resources. However,
since the label is deterministic, if the tests are run multiiple times
on the same cluster and if the resources weren't properly cleaned up
after the last test run due to an unexpected failure, the subsequent
test runs would include the old resources in their results. To prevent
this, use a unique ID for the resource label in the steveapi integration
tests.