Update service account object with attached secret
Without this patch, if the first secret stored in the service account
object is invalid, the impersonation module will time out waiting for it
to become valid. This change ensures that invalid secrets are discarded
from the object and only the most recent, known working one is kept. The
pointer to the service account is updated so that later references to
the secret name will be correct.

This change also ignores the "not found" error if it both fails to
update the service account and delete the accompanying secret, because
if the service account doesn't exist then the secret is probably already
deleted by virtue of its owner reference.