fix: proxy calls to aws fail with newer AWS SDK version
When upgrading the AWS SDK from version 1.44.253 to 1.44.266 this caused
the proxy calls to AWS to fail with `SignatureDoesNotMatch` error. This
was caused as a new partition was added (`aws-iso-e`) which currently
only has 1 service listed and does not contain IAM.

The logic loops around all the partitions trying to get the service and
region. As the new partition is last then the value for service is
overwritten with a blank string and the signing fails as the service
isn't included in the signature.

This change checks if a partition returns a service that is global. If
it does then it stops processing partitions and the existing logic then
takes over for global services.

Signed-off-by: Richard Case <richard.case@suse.com>