refactor: compute hash and owner checks
This updates the logic of the migration to compute the hash for the
machine template. This is to stop a roll out of the machine.

NOTE: if we where using CAPI 1.4.0 or higher then the computing of
hashes wouldn't be needed as labels/annotation are propogated without
doing a rollout.

Signed-off-by: Richard Case <richard.case@suse.com>