Pin Fleet version
This deprecates environment variable `CATTLE_FLEET_MIN_VERSION` in favor
of `CATTLE_FLEET_VERSION`, to enable an exact Fleet version to be
specified at Rancher installation time.

If `CATTLE_FLEET_MIN_VERSION` is set, its value will be used over
`CATTLE_FLEET_VERSION`. Regardless of the setting from which it comes,
the Fleet version chosen for installation will be installed as a min
version, to be updated only manually.

To this end, the system charts manager is updated to consider a "min
version" install as targeting that specific min version, instead of the
latest available version of the chart. However, a min version install is
still different behaviour from exact version-based install, which
enforces that version even if a more recent version is already
installed. Min version-based installs consider more recent versions
valid, and Rancher will not override/downgrade them.

This allows Fleet to be updated via the Rancher UI, and for that update
to persist across catalog updates and Rancher restarts provided that the
newly installed version is at least as recent as the configured min version.