Add Github Action to verify generated code changes
and make sure that they are committed.

That includes automatic code generation by running `go generate ./...`
and also if the go.mod files are tidy.

If we configure a successful run of this Github Action as mandatory for
merging prs then there should be no need to manual verify these steps
before releasing Rancher.