Fix flakiness in quota integration tests
When resources are added or removed from the namespace default quota of
a project, this enqueues the namespace controllers to update the quota
annotation for the namespace which in turn triggers other controllers to
update the resource quotas. Sometimes the controllers run into a
conflict error while updating the annotation and need a few seconds to
retry the sync. If the integration test checks too early, the resource
quota won't have been updated yet and the test will fail. This change
adds some breathing room to the test to give all the controllers time to
update their resources before checking the result.