Update CSP billing adapter namespace
To support PAYG offering, we need to have an umbrella helm chart to
encapsulate Rancher, usage operator, and billing adapter into a single
release. Since Rancher must be using the cattle-system namespace, this
will restrict the umbrella chart to confine to the same namespace.
Furthermore, helm does not effectively support installing subcharts
into different namespaces without hardcoding it into the subcharts,
we must also need to check for the PAYG CSP billing adapter in the
same namespace as Rancher.