Everywhere an error happens, make sure it is actually logged
In a few places we were relying on errors being printed
when returned, which is only a valid assumption in standalone
mode. Rancher startup runs this as a goroutine and was happily
ignoring all errors as a result. For sanity, it's far simpler
to just print them on the spot, so that's what I've done.