Fix check for exact version in Helm index
When installing a system chart by version, Rancher now errors if the version of the
chart found in the index is not the requested one. This previously
happened only when the version to be installed would overwrite whatever
other version was already installed.

Broadening the scope of this check fixes a bug whereby a system chart,
eg. Fleet, would be installed in its latest available version if an
invalid version number was specified in environment variables (eg.
`CATTLE_FLEET_VERSION`).