Refactor the retriable ldap connection into an interface
This change is primarily for easier testing, as it allows mocking
this interface to present test data. The retry logic and its
associated complexity is more of a necessity in the unmigration
path, which needs to tolerate some degree of infrastructure failure
and keep continuing to try to run while headless.