feat: added capi label to kubeconfig secrets
The newest version of CAPI requires that secrets that contain the kubeconfigs for  clusters
must have the cluster name label. Starting with CAPI 1.5.0 if the secrets don't have this
then the CAPI cluster reconciliation loop will not "find" the secret. This change adds the
required label for new kubeconfig secrets created.

Additionally a new migration has been added to add this label to kubeconfig secrets for
RKE2 clusters that have already been created.

Signed-off-by: Richard Case <richard.case@suse.com>

wip: moved the migration

Signed-off-by: Richard Case <richard.case@suse.com>