Pin AKS,EKS and GKE operator chart versions
Issue: https://github.com/rancher/rancher/issues/42040

This change is adding `CATTLE_AKS_OPERATOR_VERSION`, `CATTLE_EKS_OPERATOR_VERSION` and `CATTLE_GKE_OPERATOR_VERSION`, to enable an exact operator version to be specified at Rancher installation time.

The system charts manager is updated to consider a "exact
version" install as targeting that specific version, instead of the
latest available version of the chart.

This allows aks,eks and gke operators to be updated via the Rancher UI, and for that update
to persist across catalog updates and Rancher restarts provided that the
newly installed version is at least as recent as the configured version.