Add charts url setting (#43530)
* Add default chart URL setting

This adds new settings to control the origin of system, partner and RKE2
charts, which can now be tested with custom versions prior to releasing
corresponding RCs to their respective repositories:
* `ChartDefaultURL` (`rancher/charts`)
* `PartnerChartDefaultURL` (`rancher/partner- charts`)
* `RKE2ChartDefaultURL` (`rancher/rke2- charts`)

These options are intended for internal use only. When in use, Rancher
issues a warning suggesting to unset them should any issues be
encountered when fetching charts.

* Group settings by repo type

This moves each default branch setting, for the partner and RKE2 charts
repositories, next to the default URL settings for the respective
repository, to make configuration easier and improve documentation.