Simplify Secret name generation
Change from generating a name outside of the way it's generated in
CreateOrUpdateSecrets to using the name that CreateOrUpdateSecrets
returns.

This fixes the underlying problem, the name of the Secret was
effectively being guessed by consumers of the code, instead, we return
the created secret, and use the `.metadata.name` and `.metadata.namespace`
to generate the returned name.

Callers are updated to consume the returned name.