[v2.9] fix: update 'promoted' ClusterRole when removing all permissions from RoleTemplate (#44765)
* reconcileRoleForProjectAccessToGlobalResource refactored

small refactor of the reconcileRoleForProjectAccessToGlobalResource used to reconcile the promoted ClusterRole

* fix test: keep create on failures

When the Get ClusterRole was failing we were returning. This was a "breaking change" and it was failing other tests.

Keeping the old behaviour for retro compatibility. Maybe this could be changed.

* changed map[string]bool to map[string]struct{}

The map is used as a Set and the value is not used. This change will make this more explicit.

added baseRule to buildRule for testing

added tests

added tests for removal of policy rules

* fixed promoted role not updating

When removing completely all the verbs from a rule this update was skipped from the reconciliation. This fix will instead check for the removed rule and remove it from the ClusterRole.

avoid creation of new ClusterRole with no verbs

fixed test for map change

* used sets instead of maps

* changed errors.Wrap to fmt.Errorf

added missing tests

return if error is NOT notFound

* implemented suggested changes

* fixed issue with concurrency in test with sync.Map