refactor(pkg/metrics): get rid of downstream cluster caches and minimize network bandwidth
This component is only used for getting a count of certain kinds in the
downstream (or local) cluster.
These changes make explicitly use of every client kind being used,
removes cache usage and implements and indirect way of getting the
objects count: using a LIST operation limited to 1 item and read the
remaining object count from the response metadata.