Add build system's values to Drone's targets
Currently, Drone doesn't run all the scripts intended to be run on
release. Because of this, images built by Drone and pushed to registries
don't have some of the environment variables needed for Rancher at run
time.

Because Drone will soon be removed in favor of GitHub Actions, this
commit simply adds the environment variables as build-args inline,
instead of fixing the Drone pipeline steps and the scripts it runs.