[v2.7] feat: added capi label to kubeconfig secrets (#42318)
* feat: added capi label to kubeconfig secrets

The newest version of CAPI requires that secrets that contain the kubeconfigs for  clusters
must have the cluster name label. Starting with CAPI 1.5.0 if the secrets don't have this
then the CAPI cluster reconciliation loop will not "find" the secret. This change adds the
required label for new kubeconfig secrets created.

Additionally a new migration has been added to add this label to kubeconfig secrets for RKE2
clusters that have already been created.

Signed-off-by: Richard Case <richard.case@suse.com>

* Add CAPI label validation to kubeConfigValid method.

This also adapts existing Test_kubeConfigValid tests to add newly introduced
labels map and adds new test cases to verify label validation is performed.

Signed-off-by: Furkat Gofurov <furkat.gofurov@suse.com>

* refactor: migration logic updated

The migration logic has been updated to match the change for the 2.8 &
2.9 branches.

Signed-off-by: Richard Case <richard.case@suse.com>

* refactor: changes based on review feedback

Signed-off-by: Richard Case <richard.case@suse.com>

* chore: revert migration after review feedback

Signed-off-by: Richard Case <richard.case@suse.com>

---------

Signed-off-by: Richard Case <richard.case@suse.com>
Signed-off-by: Furkat Gofurov <furkat.gofurov@suse.com>
Co-authored-by: Furkat Gofurov <furkat.gofurov@suse.com>