Fix RoleTemplate nonstop syncing (#45951)
Management Plane Roles were reconciled for PRTBs, CRTBs and when a
RoleTemplate is updated. This lead to a bug were Management Plane Roles
had the same permissions as the RoleTemplate.

To address this issue, the reconciliation has been modified to
only occur for PRTBs and CRTBs, as these roles are specifically created
during the reconciliation of PRTBs and CRTBs.