build: Make Dockerfiles self-contained Self-contained Dockerfiles enables the generation of comprehensive provenance metadata straight out of Docker buildx.
Changes worth noting:
- The container images no longer copy files 'from the environment' which are not
part of the source code. Anything apart from the source is either fetched or compiled
as a layer within the Dockerfile.
- Docker caching is better optimised and only the required files are being copied into
the layers.
- docker save is not supported within the context of a build. It has now been replaced
with a adhoc CLI which pulls the k3s images file, fetches the images and then export
them into a tarball using the docker archive format.

Some of the steps that are now part of the base images are duplicated in other parts of
the code. To avoid a shotgun surgery, those will be removed as a follow-up change.

Signed-off-by: Paulo Gomes <pjbgf@linux.com>