Set annotation to guest cluster VM when deleting the cluster
If a cluster is deleted, the PVCs of the workloads are not deleted in Harvester. This is because the node driver does not know why the associated VM has to be deleted, e.g. because its parameters have changed or because the cluster is deleted.

To solve the problem, the finalizer code checks whether the deletion timestamp is set for the associated cluster when the machine resource is deleted. If this is the case, a label is set on the VM in the downstream cluster which is then processed by the node driver.

The Harvester specific code has been added to the already existing `machine-provision-remove` finalizer by reason because finalizers cannot be prioritized or influenced in the order in which they are executed. Since the Harvester-specific code must be executed BEFORE the Rancher code, it is therefore not possible to ensure that a separate finalizer is executed before the Rancher `machine-provision-remove` finalizer. Therefore, the only option is to extend the existing finalizer.

Signed-off-by: Volker Theile <vtheile@suse.com>