Re-Pin the CAPI chart version
The CAPI chart chart version used to be pinned, but this is no longer the case
since the move to GHA.

Change the workflows to pin the version in the images it produces.

Also add the missing argument for this version in dev-scripts/quick, and pass
it to the rancher process started by local provisioning and integration test
scripts. Doing this will also make the GHA provisioning test to run with the
pinned version, since it uses the script. The GHA integration tests will also
have the version pinned because they use the docker image produced in previous
steps.