Add timeout when checking if a S3 bucket exists
When creating or updating a RKE1 cluster configured with a S3 bucket for
storing ETCD backups, the store functions check if the bucket exists. This was
done without a timeout, so if S3 wasn't reachable, it could take a long and
system-dependent time for the check to fail. These functions hold a lock while
doing this check, as part of a mechanism for ensuring unique cluster
names. This meant that no new cluster could be created as long as the the
request didn't fail and the lock was held.

This commit adds a 30s timeout to the check, and a setting for this timeout so
that users in laggy environments can easily increase it if needed.