Fix: Harvester Feature Flag
Fix Harvester feature flag enabling/disabling the Harvester node driver.
When the Harvester feaature flag is disabled, the Harvester node driver
needs to be disabled to, e.g. to prevent trying do download the node
driver in an airgapped environment.
When the Harvester feature flag is enabled, the node driver must be
enabled too. Should the node driver have been deleted, it must be
re-created when the Feature flag is enabled.

related-to: rancher/rancher#48920

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>