Change underscore (_) to plus (+) before checking if the tag is a valid semver
OCI registries don't support + as a tag character. Helm adopts the convention
of changing plus (+) to an underscore (_) in chart version tags when pushing to
a registry. Therefore it is required to convert underscores in tags back to +
same as Helm does when pulling from the registry.

Cherry-picked from:  98507a569
Fixes: https://github.com/rancher/rancher/issues/50022

Original issue: https://github.com/rancher/rancher/issues/48746