Enqueue role template bindings for all k8s distros (#49251)
* Enqueue role template bindings for all k8s distros

This refreshes cluster and project role template bindings when moving a
cluster to a different namespace, regardless of the Kubernetes
distribution powering that cluster.
This was previously done only for RKE clusters, but we have since
discovered that not doing it for k3s clusters would subject them to
issues where users would lose access to a cluster after it had been
moved to a different workspace.

* Distinguish use cases by role existence

Instead of focusing on RKE vs non-RKE use cases, the test driver now
validates what happens when a role exists or does not, with role
template bindings being enqueued in the latter case.