Fix invalid memory access
Fix invalied memory access when syncronizing the Harvester feature flag
and the Harvester Baremetal Container Workloads feature flag.

The update of the Harvester Baremetal Container Workload feature flag
contained an invalid memory access at the point where the new value of
that feature flag is set.

The problem was that a feature flag's value is a pointer to a boolean
that may be nil. The nil check was only performed for the Harvester
feature flag, but not for the Harvester Baremetal Container Workload
Feature flag. As a result, when syncronizing the two feature flags,
depending on whether the Harvester Baremetal Container Workload feature
flag had been set before or not, there could be an assignment to the
memory address pointed to by a nil pointer - which is invalid.

In case the Harvester Baremetal Container Workload feature flag was not
previously set, the `update.Spec.Value` and assigning to it would be an
invalid memory access.

This is fixed by assigning directly to the pointer instead of assigning
to the address poited to by the pointer, since now the pointer can be
nil or not and will always take the new value.

related-to: rancher/rancher#49303

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>