Add safe method for setting a feature value
Add a safe method for setting the value of a v3.Feature object.
The v3.Feature object stores its value as pointer to a boolean with the
possibility of that pointer being nil.
Therefore it's not trivial to set the value of a v3.Feature because of
two edgecases:

1) A new boolean may need to be allocated first
2) Setting the v3.Feature.Spec.Value property may change the v3.Feature
   object without having changed the value, if the pointer points to a
   new location in memory, which has the same content

The SetValue method avoids both cases by checking if the value really
needs updating and by taking care of the allocation.
This preserves the expected behavior out of reflect.DeepEquals() on
v3.Feature objects and provides a safe implementation for setting a
feature's value.

related-to: rancher/rancher#49303

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>