[Backport v2.11] Fix invalid memory access (#49557)
* Fix invalid memory access

Fix invalied memory access when syncronizing the Harvester feature flag
and the Harvester Baremetal Container Workloads feature flag.

The update of the Harvester Baremetal Container Workload feature flag
contained an invalid memory access at the point where the new value of
that feature flag is set.

The problem was that a feature flag's value is a pointer to a boolean
that may be nil. The nil check was only performed for the Harvester
feature flag, but not for the Harvester Baremetal Container Workload
Feature flag. As a result, when syncronizing the two feature flags,
depending on whether the Harvester Baremetal Container Workload feature
flag had been set before or not, there could be an assignment to the
memory address pointed to by a nil pointer - which is invalid.

In case the Harvester Baremetal Container Workload feature flag was not
previously set, the `update.Spec.Value` and assigning to it would be an
invalid memory access.

This is fixed by assigning directly to the pointer instead of assigning
to the address poited to by the pointer, since now the pointer can be
nil or not and will always take the new value.

related-to: rancher/rancher#49303

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* Add safe method for setting a feature value

Add a safe method for setting the value of a v3.Feature object.
The v3.Feature object stores its value as pointer to a boolean with the
possibility of that pointer being nil.
Therefore it's not trivial to set the value of a v3.Feature because of
two edgecases:

1) A new boolean may need to be allocated first
2) Setting the v3.Feature.Spec.Value property may change the v3.Feature
   object without having changed the value, if the pointer points to a
   new location in memory, which has the same content

The SetValue method avoids both cases by checking if the value really
needs updating and by taking care of the allocation.
This preserves the expected behavior out of reflect.DeepEquals() on
v3.Feature objects and provides a safe implementation for setting a
feature's value.

related-to: rancher/rancher#49303

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

---------

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>