Set the taint annotation on the customMachine and Machine for custom clusters (#49923)
When provisioning a custom K3s cluster where all roles are assigned to each node,
an issue arises if the first machine to register has a taint that is not tolerated
by the cattle-cluster-agent or other K3s addon components. In this case, no
other machine can complete the registration process because the only available node
is unschedulable due to the taint. As a result, the cluster remains with a single
tainted node, and provisioning cannot proceed.

This occurs because the taint information is not set on the corresponding CAPI
Machine object - specifically, the rke.cattle.io/taints annotation is missing.
Without this annotation, Rancher does not apply the necessary tolerations when
rendering templates for the cattle-cluster-agent. This issue does not affect
node-driver K3s clusters, where taints are properly propagated to the CAPI Machine
objects.

With this change, cattle-cluster-agent will be able to run, so the K3s cluster will
become active and accessible in Rancher UI. K3s components are still pending for
being allocated to a node.