gh-49810: fix token identification
When creating a cloud credential for Harvester, an annotation containing
the expiration date of the authentication token used in the kubeconfig
for the Harvester cluster needs to be set.
To figure out the correct expiration date, the token needs to be
identified, but this needs to account for the fact that the token has a
different naming scheme for the default admin user and additional users.

related-to: https://github.com/rancher/rancher/issues/49810

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>