tests: add unit test cases for non-default admin
Add unit test cases for identifying tokens associated with kubeconfigs
for Harvester clusters and extracting the expiration time from them,
when the token belongs to a non-default admin user.

The default admin user's token are named kubeconfig-user-XXXXX, but for
non-default users the tokens are name kubeconfig-u-XXXXXX. These test
cases exercise the code that identifies the tokens.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>