GitHub Action for restarting CI for Testing PRs
This action triggers on completed workflows of prs with the branch name
"ci_testing_*".

It retriggers the workflow if the amount of runs is less then MAX_RUNS
specified in the pr.

After each workflow ran for MAX_RUNS it generates a report extracting a
snippet of the failed CI runs.