Use a sample CRD for documenting dynamic schemas
Preivously, dynamic schemas imported some Rancher types at run-time to
generate dynamic schemas. Because we don't have access to their comments at
run-time, they could be used for the populating descriptions in the generated
CRDs.

This commit generates a sample CRD with that also imports these types with
kubebuilder, which will read the comments as descriptions, then injects the
corresponding documented properties into the dynamic schemas. The sample CRD
is only read as a file embeded in the binary to obtain the descriptions, it
doesn't get applied to any cluster.

This commit also improves descriptions for dynamic schemas.