For lists that sort by name but not namespace, save the data secondarily-sorted by namespace.
The problem with not sorting the data is that it gets returned in random order according to
namespace (within a name) when using VAI. This causes false diffs in the saved output.
So we modify the returned objects by sorting them secondarily by namespace, and don't
have false git diffs after a run of tests.