Verify data is minimally sorted when sorting only by name.
Since we're using the JSON-rendered objects as a sanity check that
the backend is returning the correct output, we need to do a separate
check when we're sorting by name but not namespace.  The call to
`slices.IsSorted` does that.

Note that `slices.IsSortedFunc` and `slices.SortedStableFunc` want
two different kinds of initial arguments for the `~[]E` argument --
couldn't figure out what to give them in common.