Fix integration tests vai:  round04: ignore insignificant differences in the saved JSON files (#50878)
* Save a baseline of zero-timed JSON files

* Stop ignoring the JSON artifacts.

* Ignore the random namespace tag in diffs.

* For lists that sort by name but not namespace, save the data secondarily-sorted by namespace.

The problem with not sorting the data is that it gets returned in random order according to
namespace (within a name) when using VAI. This causes false diffs in the saved output.
So we modify the returned objects by sorting them secondarily by namespace, and don't
have false git diffs after a run of tests.

* Verify data is minimally sorted when sorting only by name.

Since we're using the JSON-rendered objects as a sanity check that
the backend is returning the correct output, we need to do a separate
check when we're sorting by name but not namespace.  The call to
`slices.IsSorted` does that.

Note that `slices.IsSortedFunc` and `slices.SortedStableFunc` want
two different kinds of initial arguments for the `~[]E` argument --
couldn't figure out what to give them in common.

* Remove commented-out obsolete code.