[2.12] manage SUC app and Plans in Harvester local (#50948)
Haverester installs Rancher in its local cluster and relies on Rancher to install the SUC app and the Plans for managing the Kubernetes version of the local cluster. 
It’s worth noting that in this scenario, the local cluster is treated as a Rancher-provisioned RKE2 cluster rather than an imported one.