[main] Update max-failures for windows worker nodes (#52033)
Change the `max-failures` value for Windows worker plans to `-1`. This results in Windows plans being reattempted until they succeed, which matches the existing behavior of Linux workers. Maintain a `failure-threshold` of `5` to ensure that transient errors are allowed to resolve themselves before presenting an error message in the Rancher UI.